# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "PropertyPro-lite",
    "description": "Property Pro Lite is a platform where people can create and/or search properties for sale or rent."
  },
  "servers": [
    {
      "url": "https://propertypro-lite-e.herokuapp.com/api/v1"
    },
    {
      "url": "http://localhost:8082/api/v1/"
    }
  ],
...
  "paths": {
...
    "/property": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get all property adverts",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/properties"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse1"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse1"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Create a property ad.",
        "security": [
          {
            "Token": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "price": {
                    "type": "float"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "status": {
                    "type": "string",
                    "required": false
                  },
                  "purpose": {
                    "type": "string",
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/properties"
                }
              }
            }
          },
          "401": {
            "description": "You are not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse1"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse1"
                }
              }
            }
          },
          "400": {
            "description": "bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse2"
                }
              }
            }
          }
        }
      }
    },
...
    "/property/": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get all property advertisement offering a specific type of property.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "description": "property type. e.g Duplex",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/properties"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse1"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse1"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Token": {
        "type": "apiKey",
        "in": "header",
        "name": "authorisation"
      }
    },
    "schemas": {
      "properties": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "owner": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "created_on": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "price": {
            "type": "float"
          },
          "image_url": {
            "type": "string"
          },         
          "Id": {
            "type": "string"
          }
          
        }
      },
...
      "errorResponse1": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "errorResponse2": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "object"
          }
        }
      }
    }
  }
}